+++ index.html
@@ -234,8 +234,8 @@
       </section>
     </template>
 
- 
-    <script src="js/main.js"></script>
- 
+
+    <script src="js/main.js" type="module"></script>
+
 </body>
 </html>
